Refactoring scenario code to move shared methods into a llibrary.
