--- dotnetv3/IAM/Scenarios/IAMBasics/S3Wrapper.cs
@@ -1,95 +1,0 @@
-﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier:  Apache-2.0
-
-using Amazon.S3;
-using Microsoft.Extensions.FileSystemGlobbing.Internal;
-
-namespace IAMBasics;
-
-/// <summary>
-/// A class to perform S3 actions for the IAM Basics scenario.
-/// </summary>
-internal class S3Wrapper
-{
-    private IAmazonS3 _s3Service;
-    private IAmazonSecurityTokenService _stsService;
-
-    /// <summary>
-    /// Constructor for the S3Wrapper class.
-    /// </summary>
-    /// <param name="s3Service">An Amazon S3 client object.</param>
-    /// <param name="stsService">An  AWS Security Token Service (AWS STS)
-    /// client object.</param>
-    public S3Wrapper(IAmazonS3 s3Service, IAmazonSecurityTokenService stsService)
-    {
-        _s3Service = s3Service;
-        _stsService = stsService;
-    }
-
-    /// <summary>
-    /// Assumes the IAM role for the current session.
-    /// </summary>
-    /// <param name="roleSession">A string representing the current session.</param>
-    /// <param name="roleToAssume">The name of the IAM role to assume.</param>
-    /// <returns>Credentials for the newly assumed IAM role.</returns>
-    public async Task<Credentials> AssumeS3RoleAsync(string roleSession, string roleToAssume)
-    {
-        // Create the request to use with the AssumeRoleAsync call.
-        var request = new AssumeRoleRequest()
-        {
-            RoleSessionName = roleSession,
-            RoleArn = roleToAssume,
-        };
-
-        var response = await _stsService.AssumeRoleAsync(request);
-
-        return response.Credentials;
-    }
-
-    /// <summary>
-    /// Create a new Amazon S3 bucket.
-    /// </summary>
-    /// <param name="bucketName">The name for the new bucket.</param>
-    /// <returns>A Boolean value indicated whether the action completed
-    /// successfully.</returns>
-    public async Task<bool> PutBucketAsync(string bucketName)
-    {
-        var response = _s3Service.PutBucketAsync(bucketName);
-        return response.IsCompletedSuccessfully;
-    }
-
-    /// <summary>
-    /// List the buckets that are owned by the user's account.
-    /// </summary>
-    /// <returns>Async Task</returns>
-    public async Task<List<S3Bucket>?> ListMyBucketsAsync()
-    {
-        try
-        {
-            // Get the list of buckets accessible by the new user.
-            var response = await _s3Service.ListBucketsAsync();
-
-            return response.Buckets;
-        }
-        catch (AmazonS3Exception ex)
-        {
-            // Something else went wrong. Display the error message.
-            Console.WriteLine($"Error: {ex.Message}");
-            return null;
-        }
-    }
-
-    /// <summary>
-    /// Update the client objects with new client objects. This is available
-    /// because the scenario uses the methods of this class without and then
-    /// with the proper permissions to list Amazon Simple Storage Service
-    /// (Amazon S3) buckets.
-    /// </summary>
-    /// <param name="s3Service">The Amazon S3 client object.</param>
-    /// <param name="stsService">The AWS STS client object.</param>
-    public void UpdateClients(IAmazonS3 s3Service, IAmazonSecurityTokenService stsService)
-    {
-        _s3Service = s3Service;
-        _stsService = stsService;
-    }
-}
