+++ dotnetv3/IAM/Scenarios/IAMGroups/IAMGroups.csproj
@@ -9,15 +9,16 @@
 
   <ItemGroup>
     <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.7.4" />
-    <PackageReference Include="AWSSDK.IdentityManagement" Version="3.7.100.80" />
-    <PackageReference Include="AWSSDK.S3" Version="3.7.103.15" />
-    <PackageReference Include="AWSSDK.SecurityToken" Version="3.7.101.14" />
+    <PackageReference Include="AWSSDK.IdentityManagement" Version="3.7.100.84" />
+    <PackageReference Include="AWSSDK.S3" Version="3.7.103.20" />
+    <PackageReference Include="AWSSDK.SecurityToken" Version="3.7.101.18" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\Actions\IAMActions.csproj" />
+    <ProjectReference Include="..\IamScenariosCommon\IamScenariosCommon.csproj" />
   </ItemGroup>
 
 </Project>
