--- dotnetv3/IAM/Scenarios/IAMGroups/UIWrapper.cs
@@ -1,81 +1,0 @@
-﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier:  Apache-2.0
-
-namespace IAMGroups;
-
-public class UIWrapper
-{
-    public readonly string SepBar = new('-', Console.WindowWidth);
-
-    /// <summary>
-    /// Show information about the scenario.
-    /// </summary>
-    public void DisplayOverview()
-    {
-        Console.Clear();
-
-        DisplayTitle("Welcome to the IAM Groups Demo");
-        Console.WriteLine("This example application does the following:");
-        Console.WriteLine("\t1. Creates an Amazon Identity and Access Management (IAM) group.");
-        Console.WriteLine("\t2. Adds an IAM policy to the IAM group giving it full access to Amazon S3.");
-        Console.WriteLine("\t3. Creates a new IAM user.");
-        Console.WriteLine("\t4. Creates an IAM access key for the user.");
-        Console.WriteLine("\t5. Adds the user to the IAM group.");
-        Console.WriteLine("\t6. Lists the buckets on the account.");
-        Console.WriteLine("\t7. Proves that the user has full Amazon S3 access by creating a bucket.");
-        Console.WriteLine("\t8. List the buckets again to show the new bucket.");
-        Console.WriteLine("\t9. Cleans up all the resources created.");
-    }
-
-    /// <summary>
-    /// Display a message and wait until the user presses enter.
-    /// </summary>
-    public void PressEnter()
-    {
-        Console.Write("\nPress <Enter> to continue. ");
-        _ = Console.ReadLine();
-        Console.WriteLine();
-    }
-
-    /// <summary>
-    /// Pad a string with spaces to center it on the console display.
-    /// </summary>
-    /// <param name="strToCenter"></param>
-    /// <returns></returns>
-    public string CenterString(string strToCenter)
-    {
-        var padAmount = (Console.WindowWidth - strToCenter.Length) / 2;
-        var leftPad = new string(' ', padAmount);
-        return $"{leftPad}{strToCenter}";
-    }
-
-    /// <summary>
-    /// Display a line of hyphens, the centered text of the title and another
-    /// line of hyphens.
-    /// </summary>
-    /// <param name="strTitle">The string to be displayed.</param>
-    public void DisplayTitle(string strTitle)
-    {
-        Console.WriteLine(SepBar);
-        Console.WriteLine(CenterString(strTitle));
-        Console.WriteLine(SepBar);
-    }
-
-    /// <summary>
-    /// Display a countdown and wait for a number of seconds.
-    /// </summary>
-    /// <param name="numSeconds">The number of seconds to wait.</param>
-    public void WaitABit(int numSeconds, string msg)
-    {
-        Console.WriteLine(msg);
-
-        // Wait for the requested number of seconds.
-        for (int i = numSeconds; i > 0; i--)
-        {
-            System.Threading.Thread.Sleep(1000);
-            Console.Write($"{i}...");
-        }
-
-        PressEnter();
-    }
-}
