+++ dotnetv3/IAM/Scenarios/IamScenariosCommon/IamScenariosCommon.csproj
@@ -1,24 +1,24 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AWSSDK.Core" Version="3.7.105.12" />
     <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.7.4" />
     <PackageReference Include="AWSSDK.IdentityManagement" Version="3.7.100.84" />
     <PackageReference Include="AWSSDK.S3" Version="3.7.103.20" />
     <PackageReference Include="AWSSDK.SecurityToken" Version="3.7.101.18" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\Actions\IAMActions.csproj" />
-    <ProjectReference Include="..\IamScenariosCommon\IamScenariosCommon.csproj" />
   </ItemGroup>
 
 </Project>
