+++ dotnetv3/IAM/Scenarios/IamScenariosCommon/Usings.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier:  Apache-2.0
 
-// snippet-start:[IAM.dotnetv3.IAMBasics.Usings]
+// snippet-start:[IAM.dotnetv3.Scenarios.Usings]
 global using Amazon;
 global using Amazon.IdentityManagement;
 global using Amazon.S3;
@@ -9,11 +9,11 @@ global using Amazon.S3.Model;
 global using Amazon.SecurityToken;
 global using Amazon.SecurityToken.Model;
 global using IAMActions;
-global using IamScenariosCommon;
+global using Microsoft.Extensions.FileSystemGlobbing.Internal;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.Logging;
 global using Microsoft.Extensions.Logging.Console;
 global using Microsoft.Extensions.Logging.Debug;
 
-// snippet-end:[IAM.dotnetv3.IAMBasics.Usings]
+// snippet-end:[IAM.dotnetv3.Scenarios.Usings]
