+++ dotnetv3/IAM/Tests/testsettings.json
@@ -2,5 +2,6 @@
   "UserName": "iam-test-user",
   "S3PolicyName": "s3-list-buckets-policy",
   "RoleName": "test-temporary-role",
-  "AssumePolicyName": "test-trust-user"
+  "AssumePolicyName": "test-trust-user",
+  "GroupName": "test-group"
 }
